+++ nitrite-rocksdb-adapter/src/test/java/org/dizitart/no2/integration/rocksdb/NitriteStoreEventTest.java
@@ -100,8 +100,6 @@ public class NitriteStoreEventTest  {
         assertTrue(listener.committed);
         assertTrue(listener.closing);
         assertTrue(listener.closed);
-
-        db.getStore().unsubscribe(listener);
     }
 
     @Data