+++ nitrite/src/test/java/org/dizitart/no2/common/event/EventTest.java
@@ -53,6 +53,7 @@ public class EventTest {
     private Nitrite db;
     private ObjectRepository<Employee> employeeRepository;
     private SampleListenerCollection listener;
+    private String subscription;
 
     @Rule
     public Retry retry = new Retry(3);
@@ -81,7 +82,7 @@ public class EventTest {
 
         employeeRepository = db.getRepository(Employee.class);
         listener = new SampleListenerCollection();
-        employeeRepository.subscribe(listener);
+        subscription = employeeRepository.subscribe(listener);
     }
 
     @Test
@@ -158,7 +159,7 @@ public class EventTest {
 
     @Test
     public void testDeregister() {
-        employeeRepository.unsubscribe(listener);
+        employeeRepository.unsubscribe(subscription);
         Employee e = new Employee();
         e.setEmpId(1L);
         e.setAddress("abcd");
@@ -205,7 +206,7 @@ public class EventTest {
             if (!employeeRepository.isDropped()
                     && employeeRepository.isOpen()) {
                 employeeRepository.remove(ALL);
-                employeeRepository.unsubscribe(listener);
+                employeeRepository.unsubscribe(subscription);
                 employeeRepository.close();
             }
         }