+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 ## Release 4.3.1
 
+### New Changes
+
+- GraalVM support for nitrite-mvstore-adapter #995
+- Event subscription api changes
+
 ### Issue Fixes
 
 - Fix for `Document.getFields()` not returning iterable fields
 - Fix for failing tests on systems with non-ENGLISH locale #994
+- Fix for NPE in `DefaultTransactionalRepository` #1032
 
 ## Release 4.3.0 - Jul 1, 2024